dispatch: treat fd/dirfd syscall args as 32-bit (glibc w-register tops)
Debian GNU `ln a b` under the SIGSYS tier failed with ENOENT while Alpine
worked: glibc passes int arguments (dirfds, fds) in w-registers and may
leave the x-register's top half dirty, and the handler reads the full
64-bit register — so AT_FDCWD arrived garbage-topped, compared unequal in
resolve_at_host, fell into the real-dirfd branch and failed the resolve
("[cng] linkat: src unresolved"). musl sign-extends through long, which is
why Alpine never hit it. Most other syscalls dodged the bug by accident:
xlate's mismatch falls through to reissuing the original args, which the
kernel truncates itself, and the real cwd tracks the guest cwd — but that
also meant relative-path stats under glibc skipped the l2s nlink fixup.

Truncate to int at every monitor-side consumer: resolve_at_host, xlate,
the readlinkat AT_FDCWD test, cng_l2s_deny, proc_fd_path, and
l2s_fd_count's /proc path. Kernel-bound reissues need nothing (the kernel
truncates fd args itself).

-t l2stest gains l2s-dirtyfd: openat/linkat/newfstatat driven with
AT_FDCWD in the low word and garbage in the high word must behave exactly
like clean AT_FDCWD.

+++ src/monitor/l2s.c
@@ -285,7 +285,7 @@ int cng_l2s_deny(long dirfd, const char *gp) {
      * absolute/cwd-relative paths — a ".l2s" entry elsewhere in the tree
      * stays usable, and a dirfd-relative walk can only reach the store
      * through ".." (accepted, documented). */
-    if ((gp[0] == '/' || dirfd == CNG_AT_FDCWD) && cng_g_fs) {
+    if ((gp[0] == '/' || (int)dirfd == CNG_AT_FDCWD) && cng_g_fs) {
         char canon[CNG_PATH_MAX];
         if (cng_fs_abscanon(cng_g_fs, gp, canon, sizeof canon) == 0 &&
             !strncmp(canon, "/.l2s", 5) &&
@@ -670,7 +670,8 @@ static int l2s_fd_count(long fd, unsigned long *count) {
     char link[64], *p = link;
     char *end = link + sizeof link - 1;
     p += cng_strlcpy(p, "/proc/self/fd/", (size_t)(end - p) + 1);
-    put_u64(&p, end, (unsigned long long)fd, 1);
+    /* int arg: the x-register's top half may be dirty (glibc). */
+    put_u64(&p, end, (unsigned long long)(unsigned)(int)fd, 1);
     *p = '\0';
     char path[CNG_PATH_MAX];
     long n = l2s_readlink(link, path, sizeof path - 1);
